Code/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=3.15.0
PyJWT>=2.8.0
Pillow>=9.0.0
drf-nested-routers
pillow-avif-plugin